test(cli): split test/dev/integration-1 into two files
integration-1 had 47 self-contained tests (no shared beforeAll) taking
~14 min as a single E2E chunk. Split the routing/validation tests into
integration-6 so they run as a separate chunk in CI, cutting the longest
test-dev job roughly in half.
